\name{hglm.data-package}
\alias{hglm.data-package}
\alias{hglm.data}
\docType{package}
\title{
Data for The hglm Package
}
\description{
This data-only package was created for distributing data used in the examples of the hglm package.   
}
\details{
\tabular{ll}{
Package: \tab hglm.data\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2014-07-23\cr
Discussion: \tab https://r-forge.r-project.org/forum/?group_id=558\cr
BugReports: \tab https://r-forge.r-project.org/tracker/?group_id=558\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.10)\cr
}
}
\author{
Xia Shen \cr\cr
Maintainer: Xia Shen <xia.shen@slu.se>
}
\references{
Lars Ronnegard, Xia Shen and Moudud Alam (2010). \bold{hglm: A Package for Fitting Hierarchical Generalized Linear Models}. \emph{The R Journal}, \bold{2}(2), 20-28.\cr\cr

Youngjo Lee, John A Nelder and Yudi Pawitan (2006) \emph{Generalized Linear Models with Random Effect: a unified analysis via h-likelihood.} Chapman and Hall/CRC.\cr\cr

Xia Shen, Moudud Alam, Freddy Fikse and Lars Ronnegard (2013). \bold{A novel generalized ridge regression method for quantitative genetics}. \emph{Genetics}.\cr\cr

Moudud Alam, Lars Ronnegard, Xia Shen (2014). \bold{Spatial modeling in hglm}. \emph{Submitted}.
}

\keyword{package}
